+++ DynamicIsland/models/Constants.swift
@@ -1011,7 +1011,7 @@ extension Defaults.Keys {
     )
     static let nonNotchHeight = Key<CGFloat>("nonNotchHeight", default: 32)
     static let notchHeight = Key<CGFloat>("notchHeight", default: 32)
-    static let openNotchWidth = Key<CGFloat>("openNotchWidth", default: 640)
+    static let openNotchWidth = Key<CGFloat>("openNotchWidth", default: 800)
     static let closedNotchWidth = Key<CGFloat>("closedNotchWidth", default: 150)
     static let customizePhysicalNotchWidth = Key<Bool>("customizePhysicalNotchWidth", default: false)
         //static let openLastTabByDefault = Key<Bool>("openLastTabByDefault", default: false)